+++ dotnet/src/SmartStandards.Logging.AspSupport/Logging/SmartStandards/Dummy.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Logging.SmartStandards.AspSupport {
-
-  public class AspSupportDummy {
-
-  }
-
-}