+++ dotnet/src/SmartStandards.Logging.Centralized/Logging/SmartStandards/Dummy.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Logging.SmartStandards.Centralized {
-
-  public class CentraliedSupportDummy {
-
-  }
-
-}