+++ js/src/devLogger.ts
@@ -1,7 +1,10 @@
 
 export class DevLogger {
   
-  //TODO_RWE: hier ein dev-logger bauen, der in die browser-console loggt und 
-  //optional auch in eine martstandards logging-sink
+  //TODO_RWE: hier ein DevLogger/InfLogger/BizLogger bauen, der:
+
+  // * IMMER in die browser-console loggt und
+  // * optional auch in eine 'ICentralizedLoggingSink' (SmartStandards)
+  //   inkl. vorherigem Abrufen des LoggingFilters
 
 }